Clean up / consolidate types
